Split the config crate into a bin and lib.
