+++ crates/geyser-config/Cargo.toml
@@ -13,13 +13,31 @@ license = "AGPL-3.0-or-later"
 keywords = ["solana", "geyser", "web3", "holaplex", "graphql"]
 categories = ["cryptography::cryptocurrencies", "web-programming"]
 
+[features]
+default = []
+server = [
+  "actix",
+  "actix-web",
+  "async-graphql-actix-web",
+  "sea-orm",
+  "migration",
+]
+
+[[bin]]
+name = "geyser-config-graphql"
+required-features = ["server"]
+
 [dependencies]
-actix = "0.13.0"
-actix-cors = "0.6.0-beta.10"
-actix-web = "4.0.0-rc.2"
+actix = { version = "0.13.0", optional = true }
+actix-web = { version = "4.0.0-rc.2", optional = true }
 async-graphql = "3.0.36"
-async-graphql-actix-web = "3.0.30"
+async-graphql-actix-web = { version = "3.0.30", optional = true }
 futures = "0.3.21"
+
+[dependencies.indexer-core]
+package = "holaplex-indexer-core"
+version = "=0.1.0"
+path = "../core"
 
 [dependencies.sea-orm]
 version = "0.7.1"
@@ -28,7 +46,9 @@ features = [
   "runtime-actix-rustls",
   "macros",
 ]
+optional = true
 
 [dependencies.migration]
 package = "geyser-config-migration"
 path = "../geyser-config-migration"
+optional = true
